+++ UnityProject/Assets/Scripts/Items/Tool/WelderBase.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8b628306fc939cd4e8eecd335d824dba
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
